[MDH-15] refactor : waiting line 의존성 분리
