+++ src/main/java/com/mdh/devtable/waiting/infra/persistence/WaitingLine.java
@@ -1,0 +1,10 @@
+package com.mdh.devtable.waiting.infra.persistence;
+
+import java.time.LocalDateTime;
+
+public interface WaitingLine {
+
+    void save(Long shopId, Long waitingId, LocalDateTime createdDate);
+
+    int findRank(Long shopId, Long waitingId, LocalDateTime createdDate);
+}
