+++ src/main/java/com/mdh/devtable/waiting/infra/persistence/dto/WaitingInfo.java
@@ -1,0 +1,9 @@
+package com.mdh.devtable.waiting.infra.persistence.dto;
+
+import java.time.LocalDateTime;
+
+public record WaitingInfo(
+    Long waitingId,
+    LocalDateTime waitingStartTime
+) {
+}
